\name{MdL1}

\docType{data}

\alias{MdL1}

\title{Sondage de satisfaction (salle d'escalade)}

\description{
Ce fichier permet de réaliser des analyses statistiques de satisfaction.
}

\usage{data(MdL1)}

\format{Un jeu de données avec 921 lignes et 162 colonnes.
}

\source{S. Champely, E. Boutroy, CRIS, Lyon 1 University, FRANCE}

\references{S. Champely \& E. Boutroy (2012) Extending classical methods to study customer satisfaction. An application to a private indoor climng centre in France. European Association of Sport Management, Aalborg (DN), Septembre 2012.}


\keyword{datasets}
