\name{RcmdrSMPlugin-internal}
\title{Internal RcmdrSMPlugin objects}


\alias{dotGraph}
\alias{dotchartTable}
\alias{barGraph2}
\alias{pieChart2}
\alias{frequencyDistribution2}
\alias{Histogram2}
\alias{boxPlot2}
\alias{QQPlot2}
\alias{barCode}
\alias{barCode2}
\alias{signifNC}
\alias{oneWayAnova2}

\alias{twoWayTable1}
\alias{percent}

\alias{twoWayTable2}
\alias{twoWayTable3}
\alias{signifCC}

\alias{scatterPlot2}
\alias{statNN}
\alias{signifNN}
\alias{LM2}


\alias{plotLikert}
\alias{plotCor}
\alias{enConstruction}

\description{Internal RcmdrSMPlugin objects.}
\details{These are not to be called by the user.}
\keyword{internal}
