\name{RcmdrPlugin.SM-package}
\alias{RcmdrPlugin.SM-package}
\alias{RcmdrPlugin.SM}
\docType{package}
\title{
Install the sport management Rcmdr Plug-In
}
\description{
This package provides an Rcmdr "plug-in" for studying sport management data.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.SM\cr
Type: \tab Package\cr
Version: \tab 0.3.1\cr
Date: \tab 2012-12-21\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Stéphane Champely <champely@univ-lyon1.fr>

Maintainer: Stéphane Champely <champely@univ-lyon1.fr>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}

